Thank Wilson for catching that UINT is not a PG data type. We now use BIGINT instead (#33)
(cherry picked from commit 9ba98de1c2c6bbec5cee614e627398df5922885c)

+++ src/db/schemas.js
@@ -26,7 +26,7 @@ export const tradesTable = `
     user_id UUID NOT NULL REFERENCES Users(user_id),
     ticker TEXT NOT NULL,
     action TEXT NOT NULL,
-    quantity UINT NOT NULL,
+    quantity BIGINT NOT NULL,
     unit_price_executed DECIMAL(12, 2) NOT NULL
   );
 `;
